build: fix CI failures relating to vcpkg baseline and MSVC format specifiers
- Update `builtin-baseline` in vcpkg.json to a valid hash to fix dependency resolution failures.
- Fix C4477 MSVC format string errors in `http_libuv.c` by using `C_ABSTRACT_HTTP_NUM_FORMAT` for 64-bit integer lengths.
- Remove C99 `%zd` usage in `http_lsquic.c` and cast to `long` for strict C89 MSVC compliance.
- Run clang-format over the repository to ensure strict stylistic adherence.

+++ src/http_lsquic.c
@@ -68,7 +68,7 @@ static void lsq_on_read(lsquic_stream_t *s, lsquic_stream_ctx_t *h) {
     rctx->is_complete = 1;
     lsquic_stream_close(s);
   } else {
-    LOG_DEBUG("lsq_on_read: Error nr=%zd", nr);
+    LOG_DEBUG("lsq_on_read: Error nr=%ld", (long)nr);
     rctx->error_code = EIO;
     rctx->is_complete = 1;
     lsquic_stream_close(s);
